# 8316918: Optimize conversions duplicated across phi nodes