@TobiHartmann 2023-10-04
Thanks for the contribution! I didn't look at the changes yet but executed some quick testing. 

`compiler/c2/irTests/TestPhiDuplicatedConversion.java` fails with `-XX:UseAVX=0`:

```
Failed IR Rules (2) of Methods (2)
----------------------------------
1) Method "public static short compiler.c2.irTests.TestPhiDuplicatedConversion.float2HalfFloat(boolean,float,float)" - [Failed IR rules: 1]:
   * @IR rule 1: "@compiler.lib.ir_framework.IR(applyIfCPUFeatureAnd={}, phase={DEFAULT}, applyIf={}, applyIfCPUFeatureOr={}, applyIfCPUFeature={}, counts={"_#CONV#_", "1"}, applyIfAnd={}, failOn={}, applyIfOr={}, applyIfNot={})"
     > Phase "PrintIdeal":
       - counts: Graph contains wrong number of nodes:
         * Constraint 1: "(\\d+(\\s){2}(Conv.*)+(\\s){2}===.*)"
           - Failed comparison: [found] 0 = 1 [given]
           - No nodes matched!

2) Method "public static float compiler.c2.irTests.TestPhiDuplicatedConversion.halfFloat2Float(boolean,short,short)" - [Failed IR rules: 1]:
   * @IR rule 1: "@compiler.lib.ir_framework.IR(applyIfCPUFeatureAnd={}, phase={DEFAULT}, applyIf={}, applyIfCPUFeatureOr={}, applyIfCPUFeature={}, counts={"_#CONV#_", "1"}, applyIfAnd={}, failOn={}, applyIfOr={}, applyIfNot={})"
     > Phase "PrintIdeal":
       - counts: Graph contains wrong number of nodes:
         * Constraint 1: "(\\d+(\\s){2}(Conv.*)+(\\s){2}===.*)"
           - Failed comparison: [found] 0 = 1 [given]
           - No nodes matched!
```