@jaskarth 2023-10-04
Ah, that seems to happen as the float16 intrinsics are only applied with `UseAVX>=1`, after making sure those IR rules only apply when the intrinsics are active the test passes successfully with UseAVX=0. Thank you for the testing!